Resolve a multiple definition
I moved the definition of `struct termios trm` to keyboard.c and made an
extern statement in keyboard.h to avoid every module built against
kayboard.h having its own definition of trm. It was causing build
problems as written and isn't good practice anyway. This way, trm is
global as it seems was expected before this change.

Signed-off-by: devfonks <nkeck72@netzero.net>